Extract getOperationTypes
